The order a lookup was written in is the order it keeps, and nothing sorts it
REAUDIT47's ruling. Findings 1 and 2 close by deletion: canonical ordering
was never the design, and it is the cause of both. Nothing needed a total
order — equality answers «is», the duplicate refusal, cutoff, «old»,
«changes» and indexing, while iteration and display want determinism, which
insertion order already gives. What an order costs is placing every kind
against every other for ever, and the first attempt paid it by ordering
unknown values on their printed text, which seated unequal keys together and
equal keys apart. The comparer goes, and with it that fallback, the
unimplemented null rule, its missing memo, and the error-by-message-only
hole. Duplicates are found by asking the language's own equality of the keys
already taken, and equal lookups are equal without being alike to walk —
the trade §6 named, now written down with the graph that shows it.

What may be a key is answered one level out: «Admit» refuses any value whose
kind the runtime does not know, in every position rather than in keys alone,
because refusing it only as a key leaves it legal inside a list that is then
used as one. A host integer is refused with the rest — the language's number
is a double, and an integer beside one compares unequal to it. A bare null
is not a value the language has, so it is a fault rather than an error.

And a lookup miss gives nothing, not an error: «m @ k» is «optional V», so a
forgotten miss is a compile-time error rather than a runtime one and a
«match» stays exhaustive by ordinary typing. A list index out of range stays
an error, because a missing key is data and an index past the end is a
mistake. The spec was right and the handoff was wrong, so the code moved.

+++ Test/Unit/Indexing.cs
@@ -188,7 +188,7 @@ public class Indexing
 
         for (var tick = 1; tick <= 3; ++tick)
         {
-            graph.Write("tick", tick);
+            graph.Write("tick", (double)tick);
             graph.Step();
             Assert.Equal(1d, graph.Read("watching"));
         }
